requirement.txt
aliyun-python-sdk-core-v3==2.13.10
et-xmlfile==1.0.1
jdcal==1.4.1
jmespath==0.9.4
numpy==1.16.4
opencv-python==4.1.0.25
openpyxl==2.6.2
Pillow==6.1.0
PyQt5==5.10.0
